Guardrails/pyproject.toml
[project]
name = "guardrails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
